Update demo.html
Moved FOIA request display code to separate JavaScript file

Previously, the code for displaying FOIA requests from the MuckRock API was included as a script block in an HTML file. This commit moves the relevant code to a separate JavaScript file for better organization and modularity.

Specifically, the following code was moved from the HTML file:

- The variable declarations for 'username', 'allowedOrigins', 'corsProxyUrl', and 'apiUrl'
- The function 'isAllowedOrigin'
- The code for sending a request to the MuckRock API using the 'fetch' function and a CORS proxy, processing the response data, and adding it to the HTML page.

No functionality has been changed in this commit. The HTML file now includes a reference to the new JavaScript file, and the code is called from there instead.

This commit should make it easier to maintain and update the FOIA request display feature going forward.