Create muck.js
Add new file muck.js for MuckRock API integration

This commit adds a new JavaScript file called `muck.js`, which will contain code for integrating with the MuckRock API to display FOIA requests on a web page. The file is currently empty, but will be populated with relevant code in future commits.

This change was made in order to separate concerns and improve organization in the project, as well as to make it easier to maintain and update the MuckRock API integration feature going forward.

No functionality has been changed in this commit, and the existing HTML files do not reference the new `muck.js` file yet. Future commits will update the HTML files accordingly.