Update demo.html
Add MuckRock API integration to website using muck.js

This commit adds a new JavaScript file called `muck.js`, which contains code for integrating with the MuckRock API to display FOIA requests on a web page. The file is embedded in the HTML file using a `script` tag with the `src` attribute set to `muck.js`.

In addition, the relevant HTML file has been updated to call functions and use variables defined in `muck.js` to display the FOIA request data on the page. Specifically, the `requestList` element now references a variable defined in `muck.js`, and the `fetchMuckRequests` function is called on page load to fetch and display the requests.

This change was made in order to separate concerns and improve organization in the project, as well as to make it easier to maintain and update the MuckRock API integration feature going forward.

No functionality has been changed in this commit beyond the addition of the `muck.js` file and its associated changes to the HTML file. The code in `muck.js` is currently hardcoded to display requests for a specific user, but future commits will make this more configurable.